Refine the wordlist matching filter.
It now uses regex group back references to only include words that could
possibly match, given multiple occurences of characters in the
ciphertext.

So, given a ciphertext like `JVXVL`, we will only match five-letter
words that have the same letter in the second and fourth positions, such
as 'LEVER' or 'MOTOR'.