Add features for finding the key.
This lists all the decimations of the current chain, with letters V W X
Y Z highlighted, and allows you to choose a particular decimation step
to examine.

Once a step is chosen, the display shows each of the possible matrices
for hat widths 7 through 12, with highlighted letters where at least two
sequential letters in a column match a subsequence of the chain.