Clean up the CLI arguments and populate README.
The -i/--interactive argument didn't make much sense since the program
only works interactively, so just remove it.

Requiring puzzles to be located in 'puzzles/' under the script dir was a
bit overly prescriptive. When I tried explaining that in the arguments
help text and in the README it was awkward so I ditched it. Now you can
put the puzzle file wherever you like and just give the path to it as
the first positional argument to the script.