Allow setting the hat manually.
Sometimes the hat will not be in the word list at all, in which case it
makes sense to allow the user to input a hat by hand and have it checked
off against the letter sequence.

I also made it so the HatView calculates the list of matches once when
you run it, and only re-calculates when you update the filter. And it
now shows a 'loading' page instead of just remaining blank while it
calculates.